+++ Sources/Player/Player/Player+ItemNavigation.swift
@@ -7,31 +7,63 @@
 import AVFoundation
 import DequeModule
 
-extension Player {
+public extension Player {
+    /// Checks whether returning to the previous content is possible.
+    ///
+    /// - Returns: `true` if possible.
+    ///
+    /// The behavior of this method is adjusted to wrap around both ends of the item queue when ``Player/repeatMode``
+    /// has been set to ``RepeatMode/all``.
+    ///
+    /// > Important: Observes the ``PlayerConfiguration/navigationMode`` set in the ``Player/configuration``.
     func canReturnToPreviousItem() -> Bool {
-        canReturnToItem(before: currentItem, in: storedItems)
+        canReturnToItem(before: currentItem, in: storedItems, streamType: streamType)
     }
 
+    /// Returns to the previous content.
+    ///
+    /// The behavior of this method is adjusted to wrap around both ends of the item queue when ``Player/repeatMode``
+    /// has been set to ``RepeatMode/all``.
+    ///
+    /// > Important: Observes the ``PlayerConfiguration/navigationMode`` set in the ``Player/configuration``.
     func returnToPreviousItem() {
-        guard let index = index(before: currentItem, in: storedItems) else { return }
-        queuePlayer.replaceItems(
-            with: AVPlayerItem.playerItems(
-                from: Array(storedItems),
-                after: index,
-                repeatMode: repeatMode,
-                length: configuration.preloadedItems,
-                reload: true,
-                configuration: configuration,
-                resumeState: resumeState,
-                limits: limits
+        if shouldSeekToStartTime() {
+            seek(near(.zero))
+        }
+        else if let index = index(before: currentItem, in: storedItems) {
+            queuePlayer.replaceItems(
+                with: AVPlayerItem.playerItems(
+                    from: Array(storedItems),
+                    after: index,
+                    repeatMode: repeatMode,
+                    length: configuration.preloadedItems,
+                    reload: true,
+                    configuration: configuration,
+                    resumeState: resumeState,
+                    limits: limits
+                )
             )
-        )
+        }
     }
 
+    /// Checks whether moving to the next content is possible.
+    ///
+    /// - Returns: `true` if possible.
+    ///
+    /// The behavior of this method is adjusted to wrap around both ends of the item queue when ``Player/repeatMode``
+    /// has been set to ``RepeatMode/all``.
+    ///
+    /// > Important: Observes the ``PlayerConfiguration/navigationMode`` set in the ``Player/configuration``.
     func canAdvanceToNextItem() -> Bool {
         canAdvanceToItem(after: currentItem, in: storedItems)
     }
 
+    /// Moves to the next content.
+    ///
+    /// The behavior of this method is adjusted to wrap around both ends of the item queue when ``Player/repeatMode``
+    /// has been set to ``RepeatMode/all``.
+    ///
+    /// > Important: Observes the ``PlayerConfiguration/navigationMode`` set in the ``Player/configuration``.
     func advanceToNextItem() {
         guard let index = index(after: currentItem, in: storedItems) else { return }
         queuePlayer.replaceItems(
@@ -50,14 +82,37 @@ extension Player {
 }
 
 extension Player {
-    func canReturnToItem(before item: PlayerItem?, in items: Deque<PlayerItem>) -> Bool {
-        index(before: item, in: items) != nil
+    func canReturnToItem(before item: PlayerItem?, in items: Deque<PlayerItem>, streamType: StreamType) -> Bool {
+        switch configuration.navigationMode {
+        case .smart where streamType == .onDemand:
+            return true
+        default:
+            return index(before: item, in: items) != nil
+        }
     }
 
     func canAdvanceToItem(after item: PlayerItem?, in items: Deque<PlayerItem>) -> Bool {
         index(after: item, in: items) != nil
     }
+}
 
+private extension Player {
+    func isAwayFromStartTime(withInterval interval: TimeInterval) -> Bool {
+        let time = time()
+        return time.isValid && seekableTimeRange.isValid && (time - seekableTimeRange.start).seconds >= interval
+    }
+
+    func shouldSeekToStartTime() -> Bool {
+        switch configuration.navigationMode {
+        case .immediate:
+            return false
+        case let .smart(interval: interval):
+            return (streamType == .onDemand && isAwayFromStartTime(withInterval: interval)) || !canReturnToPreviousItem()
+        }
+    }
+}
+
+extension Player {
     func replaceCurrentItemWithItem(_ item: PlayerItem?) {
         if let item {
             if let index = storedItems.firstIndex(of: item) {
